Remove viper implementation of balance module
